[Refactor] restructure project directories
