+++ src/main/java/com/capston/favicon/user/application/service/AuthService.java
