+++ src/main/java/com/capston/favicon/user/domain/Data.java
