+++ src/main/java/com/capston/favicon/user/domain/User.java
